less: don't use -X on versions of less after v530
less 530 and onward doesn't require -X when the content fits in one
screen to print it on the scrollback screen. It means that when the
file occupies more than one screen, less goes into alternate screen
mode and the file contents aren't printed on the scrollback screen.
It also means that, if the file doesn't fit in one screen, after
less exits, the file contents won't be available in the scrollback
screen.